requirements.txt
beautifulsoup4==4.13.3
botcity-framework-base==0.4.4
botcity-framework-web==0.10.0
botcity-maestro-sdk==0.6.0
certifi==2025.1.31
charset-normalizer==3.4.1
distro==1.9.0
idna==3.10
importlib_metadata==8.6.1
msedge-selenium-tools==3.141.3
numpy==1.26.4
opencv-python==4.11.0.86
packaging==24.2
pillow==11.1.0
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
selenium==3.141.0
soupsieve==2.6
typing_extensions==4.12.2
urllib3==1.26.20
webdriver-manager==4.0.2
zipp==3.21.0
